Remove type casting from positional encoding

+++ supervised_learning/0x11-attention/10-transformer_decoder.py
@@ -120,8 +120,7 @@ class Decoder(tf.keras.layers.Layer):
 
         x = self.embedding(x)
         x *= tf.math.sqrt(tf.cast(self.dm, tf.float32))
-        x += tf.cast(self.positional_encoding,
-                     dtype=tf.float32)[:seq_len]
+        x += self.positional_encoding[:seq_len]
 
         x = self.dropout(x, training=training)
